add About Us page, add to routing, update nav
This change does remove some changes made in #185 - the entire Footer links section can't just open in a new window because the logic ("this is an external site") isn't true anymore. So we'll have to either stop doing that OR refactor how the links are stored as an environment variable to pass along an additional value telling the header component how to handle them.