Allow code to be hot-reloaded with Docker
We now mount in the source code instead of copying it into the image: this allows us to propagate changes in our host file system to the running Docker container.

Test as follows:

```
$ docker-compose down -v
$ docker-compose up --build
```

And then try changing a random api url path while the containing is running. It should reflect immediately upon reload.